requirements.txt
-umpy @ file:///D:/bld/numpy_1604946136447/work
absl-py==0.9.0
anykeystore==0.2
apex==0.1
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
audioread==2.1.8
axial-positional-embedding==0.2.1
backcall==0.2.0
bleach==1.5.0
blurhash==1.1.4
boost==0.1
boto==2.48.0
cached-property==1.3.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.10.0
chardet==3.0.4
cloudpickle==1.6.0
cmake==3.18.4.post1
cntk==2.7
colorama==0.4.3
contrastive-learner==0.1.0
cryptacular==1.5.5
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dlib @ file:///D:/bld/dlib_1602513840626/work
dominate==2.6.0
easydict==1.6
entrypoints==0.2.3
face-alignment==1.1.0
fire==0.3.1
freetype-py==2.2.0
future==0.18.2
gast==0.3.3
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.30.0
h5py==2.10.0
html5lib==0.9999999
hupper==1.10.2
idna==2.6
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
imutils==0.5.3
iniconfig @ file:///home/conda/feedstock_root/build_artifacts/iniconfig_1596221474428/work
iohandler==1.0.6
ipykernel==4.6.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.17.2
Jinja2==2.9.6
joblib==0.16.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.6.0
linear-attention-transformer==0.14.1
linformer==0.2.0
llvmlite==0.31.0
lmdb==1.0.0
local-attention==1.0.2
Markdown==3.2.2
MarkupSafe==1.0
Mastodon.py==1.5.1
matplotlib==3.3.0
mistune==0.7.4
mock==4.0.2
moviepy==1.0.3
msgpack==1.0.0
mxnet-cu90==1.5.0
nbconvert==5.2.1
nbformat==4.4.0
networkx==2.4
neuralgym @ git+https://github.com/JiahuiYu/neuralgym@88292adb524186693a32404c0cfdc790426ea441
nose==1.3.7
notebook==5.0.0
numba==0.48.0
numpy==1.16.0
oauthlib==3.1.0
olefile==0.44
opencv-contrib-python==4.3.0.36
opencv-python==4.3.0.36
openmesh==1.1.4.post18
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
PasteDeploy==2.1.1
path.py==10.3.1
pathlib==1.0.1
pbkdf2==1.3
pickleshare==0.7.5
Pillow==5.4.1
plaster==1.0
plaster-pastedeploy==0.7
pooch==1.1.1
product-key-memory==0.1.10
proglog==0.1.9
progressbar2==3.53.1
prompt-toolkit==3.0.5
protobuf==3.12.4
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.18
pydot-ng==2.0.0
pyglet==1.5.7
Pygments==2.6.1
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyramid==1.10.5
pyramid-mailer==0.15.1
pyreadline==2.1
pyrender==0.1.43
python-dateutil==2.8.1
python-magic==0.4.18
python-utils==2.4.0
python3-openid==3.2.0
pytorch-fid==0.1.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==16.0.2
qtconsole==4.3.1
repoze.sendmail==4.4.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
resnest @ git+https://github.com/zhanghang1989/ResNeSt@5fe47e93bd7e098d15bc278d8ab4812b82b49414
retry==0.9.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-video==1.1.11
scipy==1.2.1
simplegeneric==0.8.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.19
structlog==20.1.0
tensorboard==1.14.0
tensorboardX==1.4
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-gpu-estimator==2.2.0
tensorlayer==2.2.3
termcolor==1.1.0
testpath==0.3
threadpoolctl==2.1.0
tifffile==2020.7.24
torch @ file:///D:/downloads/torch-1.2.0-cp36-cp36m-win_amd64.whl
torchvision @ file:///D:/downloads/torchvision-0.4.0-cp36-cp36m-win_amd64.whl
tornado==4.5.2
tqdm==4.48.2
traitlets==4.3.3
transaction==3.0.1
translationstring==1.4
trimesh==3.8.6
typing==3.7.4.3
urllib3==1.22
vector-quantize-pytorch==0.1.0
velruse==1.1.1
venusian==3.0.0
wcwidth==0.2.5
WebOb==1.8.6
Werkzeug==1.0.1
widgetsnbextension==3.0.2
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
yacs==0.1.8
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.2.0
zope.sqlalchemy==1.3
